+++ src/hkimjp/konpy2/answers.clj
@@ -56,7 +56,10 @@
           [:input {:type "hidden" :name "to" :value e}]
           [:input {:type "hidden" :name "author" :value author}]
           [:input {:type "hidden" :name "pid" :value p}]
-          [:textarea.border-1.p-2 {:class "w-2/3" :name "comment"}]
+          [:textarea
+           {:class "bg-green-100 h-40 border-1 p-2 w-2/3"
+            :name "comment"
+            :placeholder "markdown OK"}]
           (for [pt ["A" "B" "C"]]
             [:button {:class btn :name "pt" :value pt} pt])]])))